Process bot commands!
We check if a message that was broadcasted is a bot command.
If it is, and a valid one, we act on it.
If an error comes up, we boradcast it too.

Note on the tests, they're not ideal.
I sleep the tests for the goroutines to be processed.
Ideally, wait groups and the like are better solutions.
For now, I do get the assurance I need to keep developing